oc-cluster-up: create /etc/docker/daemon.json right away
Since docker-ce v23 appeared in the docker rpm repo
our oc-cluster-up role started to fail with
Error, could not touch target: [Errno 2] No such file or directory: '/etc/docker/daemon.json'

Previously we
1. started docker service
2. touched /etc/docker/daemon.json
3. added insecure_registry into it
4. restarted the docker service

I don't remember why there was the 1. step, probably because it created the
/etc/docker/. With docker v23 the 2. step fails probably because the
/etc/docker/ is not created in the 1. step.
So let's create it right away and start the docker service afterwards.

Now the steps are:
1. Add insecure_registry to /etc/docker/daemon.json (create if not exist)
3. start docker service

Unlike the file module, the lineinfile module doesn't mind if the
parent directory doesn't exist, it creates it as well.